imx8mp: add memory size detection
SolidRun supports Samsung 1GByte at 4Gtps and Micron 3GByte at 3.2Gtps
speeds.
This code tries to train using Micron memory first and then fallbacks to
Samsung in case the training failed, or memory mirror test failed while
testing in multiplications of 1GByte.

When 2GByte memory support will be added this code needs to be
revisited.

In the background we are trying to get an lpddr4 mode register read functional
which will enable proper detection of the memory vendor, size and
parameters.

Signed-off-by: Rabeeh Khoury <rabeeh@solid-run.com>